% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{augment.gamFit}
\alias{augment.gamFit}
\title{Augment function for class \code{gamFit}}
\usage{
\method{augment}{gamFit}(x, ...)
}
\arguments{
\item{x}{A \code{gamFit} object.}

\item{...}{Other arguments not currently used.}
}
\value{
A \code{tibble}.
}
\description{
Generates residuals and fitted values of a fitted \code{gamFit} object.
}
