% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{augment.lmFit}
\alias{augment.lmFit}
\title{Augment function for class \code{lmFit}}
\usage{
\method{augment}{lmFit}(x, ...)
}
\arguments{
\item{x}{A \code{lmFit} object.}

\item{...}{Other arguments not currently used.}
}
\value{
A \code{tibble}.
}
\description{
Generates residuals and fitted values of a fitted \code{lmFit} object.
}
