% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{augment.pprFit}
\alias{augment.pprFit}
\title{Augment function for class \code{pprFit}}
\usage{
\method{augment}{pprFit}(x, ...)
}
\arguments{
\item{x}{A \code{pprFit} object.}

\item{...}{Other arguments not currently used.}
}
\description{
Generates residuals and fitted values of a fitted \code{pprFit} object.
}
