% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{augment.smimodel}
\alias{augment.smimodel}
\title{Augment function for class \code{smimodel}}
\usage{
\method{augment}{smimodel}(x, data, neighbour = 0, ...)
}
\arguments{
\item{x}{A \code{smimodel} object.}

\item{data}{Training data set on which models will be trained. Must be a data
set of class\code{tsibble}.}

\item{neighbour}{If multiple models are fitted: Number of neighbours of each
key (i.e. grouping variable) to be considered in model fitting to handle
smoothing over the key. Should be an integer. If \code{neighbour = x}, \code{x}
number of keys before the key of interest and \code{x} number of keys after the
key of interest are grouped together for model fitting. The default is \code{0}
(i.e. no neighbours are considered for model fitting).}

\item{...}{Other arguments not currently used.}
}
\description{
Generates residuals and fitted values of a fitted \code{smimodel} object.
}
