% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{augment.smimodelFit}
\alias{augment.smimodelFit}
\title{Augment function for class \code{smimodelFit}}
\usage{
\method{augment}{smimodelFit}(x, ...)
}
\arguments{
\item{x}{A \code{smimodelFit} object.}

\item{...}{Other arguments not currently used.}
}
\description{
Generates residuals and fitted values of a fitted \code{smimodelFit} object.
}
