% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.smimodel}
\alias{autoplot.smimodel}
\title{Plot estimated smooths from a fitted \code{smimodel}}
\usage{
\method{autoplot}{smimodel}(object, model = 1, ...)
}
\arguments{
\item{object}{A \code{smimodel} object.}

\item{model}{An \code{integer} to indicate the smooths of which model (out of
the set of multiple models fitted) to be plotted.}

\item{...}{Other arguments not currently used.}
}
\value{
Plot(s) of fitted spline(s).
}
\description{
Plots the graphs of fitted spline(s). If a set of multiple models are fitted,
plots graphs of fitted spline(s) of a specified model (in argument
\code{model}) out of the set of multiple models fitted.
}
