% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb.R
\name{avgWidth}
\alias{avgWidth}
\title{Calculate interval forecast width}
\usage{
avgWidth(
  object,
  level = 95,
  includemedian = FALSE,
  window = NULL,
  na.rm = FALSE
)
}
\arguments{
\item{object}{An object of class  \code{bb_cvforecast}.}

\item{level}{Target confidence level for prediction intervals.}

\item{includemedian}{If \code{TRUE}, the median interval width will also be returned.}

\item{window}{If not \code{NULL}, the rolling mean (and rolling median if applicable)
matrix for interval width will also be returned.}

\item{na.rm}{A logical indicating whether \code{NA} values should be stripped
before the rolling mean and rolling median computation proceeds.}
}
\value{
A list of class \code{"width"} with the following components:
\item{width}{Forecast interval width as a multivariate time series, where the \eqn{h}th
column holds the interval width for the forecast horizon \eqn{h}. The time index
corresponds to the period for which the forecast is produced.}
\item{mean}{Mean interval width across the validation set.}
\item{rollmean}{If \code{window} is not NULL, a matrix of the rolling means
of interval width will be returned.}
\item{median}{Median interval width across the validation set.}
\item{rollmedian}{If \code{window} is not NULL, a matrix of the rolling medians
of interval width will be returned.}
}
\description{
This is a wrapper for the function \code{conformalForecast::width()}.
Calculates the mean width of prediction intervals on the validation set.
If \code{window} is not \code{NULL}, a matrix of the rolling means of interval
width is also returned. If \code{includemedian} is \code{TRUE},
the information of the median interval width will be returned.
}
