% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward.R
\name{backward}
\alias{backward}
\title{Nonparametric Additive Model with Backward Elimination}
\usage{
backward(
  data,
  val.data,
  yvar,
  family = gaussian(),
  log.transformed = FALSE,
  neighbour = 0,
  s.vars = NULL,
  s.basedim = NULL,
  linear.vars = NULL,
  tol = 0.001,
  parallel = FALSE,
  workers = NULL
)
}
\arguments{
\item{data}{The data set on which the model(s) will be trained. Must be a
data set of class \code{tsibble}.}

\item{val.data}{Validation data set. (The data set on which the model
selection will be performed.) Must be a data set of class \code{tsibble}.}

\item{yvar}{Name of the response variable as a character string.}

\item{family}{A description of the error distribution and link function to be
used in the model (see \code{glm} and \code{family}).}

\item{log.transformed}{Whether the response is log-transformed or not.}

\item{neighbour}{If multiple models are fitted: Number of neighbours of each
key (i.e. grouping variable) to be considered in model fitting to handle
smoothing over the key. Should be an integer. If \code{neighbour = x}, \code{x}
number of keys before the key of interest and \code{x} number of keys after the
key of interest are grouped together for model fitting. The default is \code{0}
(i.e. no neighbours are considered for model fitting).}

\item{s.vars}{A character vector of names of the predictor variables for
which splines should be fitted (i.e. non-linear predictors). (default:
NULL)}

\item{s.basedim}{Dimension of the bases used to represent the smooth terms
corresponding to \code{s.vars}. (For more information refer \code{mgcv::s()}.)
(default: NULL)}

\item{linear.vars}{A character vector of names of the predictor variables
that should be included linearly into the model (i.e. linear predictors).
(default: NULL)}

\item{tol}{Tolerance for the ratio of relative change in MAPE, used in model
selection. (default: 0.001)}

\item{parallel}{Whether to use parallel computing in model selection or not.
(default: FALSE)}

\item{workers}{If \code{parallel = TRUE}, number of workers to use. (default:
NULL)}
}
\description{
Fits a nonparametric additive model, with simultaneous variable selection
through a backward elimination procedure as proposed by Fan and Hyndman
(2012).
}
\details{
Fits a nonparametric multiple index model to the data, with simultaneous
variable selection (hence "sparse").
}
\references{
Fan, S. & Hyndman, R.J. (2012). Short-Term Load Forecasting Based
on a Semi-Parametric Additive Model. \emph{IEEE Transactions on Power Systems},
27(1), 134-141. \url{http://doi.org/10.1109/TPWRS.2011.2162082}
}
