% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_bootstrap.R
\name{crossVal_bb}
\alias{crossVal_bb}
\title{Single season block bootstrap prediction intervals - to be corrected}
\usage{
crossVal_bb(
  data,
  newdata,
  yvar,
  neighbour = 0,
  predictor.vars,
  modelfun,
  h = 1,
  season.period = 1,
  m = 1,
  num.futures = 1000,
  level = c(80, 95),
  forward = TRUE,
  initial = 1,
  window = NULL,
  recursive = FALSE,
  recursive_colNames = NULL,
  ...
)
}
\arguments{
\item{data}{Training data set. Must be a data set of class \code{tsibble}.(Make
sure there are no additional date/time/date-time/yearmonth/POSIXct/POSIXt
variables except for the \code{index} of the \code{tsibble}). If multiple models are
fitted, the grouping variable should be the key of the \code{tsibble}.}

\item{newdata}{Test data set. Must be a data set of class \code{tsibble}.}

\item{yvar}{Name of the response variable as a character string.}

\item{neighbour}{If multiple models are fitted: Number of neighbours of each
key (i.e. grouping variable) to be considered in model fitting to handle
smoothing over the key. Should be an integer. If \code{neighbour = x}, \code{x}
number of keys before the key of interest and \code{x} number of keys after the
key of interest are grouped together for model fitting. The default is \code{0}
(i.e. no neighbours are considered for model fitting).}

\item{predictor.vars}{A character vector of names of the predictor variables.}

\item{modelfun}{Model fitting function.}

\item{h}{Forecast horizon.}

\item{season.period}{Length of the seasonal period.}

\item{m}{Multiplier. (Block size = \code{season.period} * \code{m})}

\item{num.futures}{Number of possible future sample paths to be generated.}

\item{level}{Confidence level for prediction intervals.}

\item{forward}{If \code{TRUE}, the final forecast origin for forecasting is
\eqn{y_T}. Otherwise, the final forecast origin is \eqn{y_{T-1}}.}

\item{initial}{Initial period of the time series where no cross-validation
forecasting is performed.}

\item{window}{Length of the rolling window. If \code{NULL}, a rolling window
will not be used.}

\item{recursive}{Whether to obtain recursive forecasts or not (default -
FALSE).}

\item{recursive_colNames}{If \code{recursive = TRUE}, a character vector giving
the names of the columns in test data to be filled with forecasts.}

\item{...}{Other arguments not currently used.}
}
\description{
Compute prediction intervals by applying the single season block bootstrap
method.
}
