% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast.backward}
\alias{forecast.backward}
\title{Forecasting using nonparametric additive models with backward elimination}
\usage{
\method{forecast}{backward}(
  object,
  h = 1,
  level = c(80, 95),
  newdata,
  exclude.trunc = NULL,
  recursive = FALSE,
  recursive_colRange = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{backward}. Usually the result of a
call to \code{\link{model_backward}}.}

\item{h}{Forecast horizon.}

\item{level}{Confidence level for prediction intervals.}

\item{newdata}{The set of new data on for which the forecasts are required
(i.e. test set; should be a \code{tsibble}).}

\item{exclude.trunc}{The names of the predictor variables that should not be
truncated for stable predictions as a character string.}

\item{recursive}{Whether to obtain recursive forecasts or not (default -
\code{FALSE}).}

\item{recursive_colRange}{If \code{recursive = TRUE}, the range of column
numbers in \code{newdata} to be filled with forecasts.}

\item{...}{Other arguments not currently used.}
}
\value{
An object of class \code{forecast}. Here, it is a list containing the
following elements: \item{method}{The name of the forecasting method as a
character string.} \item{model}{The fitted model.} \item{mean}{Point
forecasts as a time series.} \item{residuals}{Residuals from the fitted
model.} \item{fitted}{Fitted values (one-step forecasts).}
}
\description{
Returns forecasts and other information for nonparametric additive models
with backward elimination.
}
\examples{
library(dplyr)
library(tibble)
library(tidyr)
library(tsibble)

# Simulate data
n = 1215
set.seed(123)
sim_data <- tibble(x_lag_000 = runif(n)) |>
  mutate(
    # Add x_lags
    x_lag = lag_matrix(x_lag_000, 5)) |>
  unpack(x_lag, names_sep = "_") |>
  mutate(
    # Response variable
    y = (0.9*x_lag_000 + 0.6*x_lag_001 + 0.45*x_lag_003)^3 + rnorm(n, sd = 0.1),
    # Add an index to the data set
    inddd = seq(1, n)) |>
  drop_na() |>
  select(inddd, y, starts_with("x_lag")) |>
  # Make the data set a `tsibble`
  as_tsibble(index = inddd)

# Training set
sim_train <- sim_data[1:1000, ]
# Validation set
sim_val <- sim_data[1001:1200, ]
# Test set
sim_test <- sim_data[1201:1210, ]

# Predictors taken as non-linear variables
s.vars <- colnames(sim_data)[3:8]

# Model fitting
backwardModel <- model_backward(data = sim_train,
                                val.data = sim_val,
                                yvar = "y",
                                s.vars = s.vars)
forecast(backwardModel, newdata = sim_test)

}
