% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_smimodel.R
\name{greedy.fit}
\alias{greedy.fit}
\title{Greedy search for tuning penalty parameters}
\usage{
greedy.fit(
  data,
  val.data,
  yvar,
  neighbour = 0,
  family = gaussian(),
  index.vars,
  initialise = c("ppr", "additive", "linear", "multiple", "userInput"),
  num_ind = 5,
  num_models = 5,
  seed = 123,
  index.ind = NULL,
  index.coefs = NULL,
  s.vars = NULL,
  linear.vars = NULL,
  nlambda = 100,
  lambda.min.ratio = 1e-04,
  refit = TRUE,
  M = 10,
  max.iter = 50,
  tol = 0.001,
  tolCoefs = 0.001,
  TimeLimit = Inf,
  MIPGap = 1e-04,
  NonConvex = -1,
  verbose = FALSE,
  parallel = FALSE,
  workers = NULL,
  exclude.trunc = NULL,
  recursive = FALSE,
  recursive_colRange = NULL
)
}
\arguments{
\item{data}{Training data set on which models will be trained. Must be a data
set of class \code{tsibble}.(Make sure there are no additional date or time
related variables except for the \code{index} of the \code{tsibble}).}

\item{val.data}{Validation data set. (The data set on which the penalty
parameter selection will be performed.) Must be a data set of class
\code{tsibble}. (Once the penalty parameter selection is completed, the
best model will be re-fitted for the combined data set \code{data +
  val.data}.)}

\item{yvar}{Name of the response variable as a character string.}

\item{neighbour}{\code{neighbour} argument passed from the outer function.}

\item{family}{A description of the error distribution and link function to be
used in the model (see \code{\link{glm}} and \code{\link{family}}).}

\item{index.vars}{A \code{character} vector of names of the predictor
variables for which indices should be estimated.}

\item{initialise}{The model structure with which the estimation process
should be initialised. The default is \code{"ppr"}, where the initial model
is derived from projection pursuit regression. The other options are
\code{"additive"} - nonparametric additive model, \code{"linear"} - linear
regression model (i.e. a special case single-index model, where the initial
values of the index coefficients are obtained through a linear regression),
\code{"multiple"} - multiple models are fitted starting with different
initial models (number of indices = \code{num_ind}; \code{num_models}
random instances of the model (i.e. the predictor assignment to indices and
initial index coefficients are generated randomly) are considered), and the
final optimal model with the lowest loss is returned, and
\code{"userInput"} - user specifies the initial model structure (i.e. the
number of indices and the placement of index variables among indices) and
the initial index coefficients through \code{index.ind} and
\code{index.coefs} arguments respectively.}

\item{num_ind}{If \code{initialise = "ppr"} or \code{"multiple"}: an
\code{integer} that specifies the number of indices to be used in the
model(s). The default is \code{num_ind = 5}.}

\item{num_models}{If \code{initialise = "multiple"}: an \code{integer} that
specifies the number of starting models to be checked. The default is
\code{num_models = 5}.}

\item{seed}{If \code{initialise = "multiple"}: the seed to be set when
generating random starting points.}

\item{index.ind}{If \code{initialise = "userInput"}: an \code{integer} vector
that assigns group index for each predictor in \code{index.vars}.}

\item{index.coefs}{If \code{initialise = "userInput"}: a \code{numeric}
vector of index coefficients.}

\item{s.vars}{A \code{character} vector of names of the predictor variables
for which splines should be fitted individually (rather than considering as
part of an index).}

\item{linear.vars}{A \code{character} vector of names of the predictor
variables that should be included linearly into the model.}

\item{nlambda}{The number of values for lambda0 (penalty parameter for L0
penalty) - default is 100.}

\item{lambda.min.ratio}{Smallest value for lambda0, as a fraction of
lambda0.max (data derived).}

\item{refit}{Whether to refit the model combining training and validation
sets after parameter tuning. If \code{FALSE}, the final model will be
estimated only on the training set.}

\item{M}{Big-M value used in MIP.}

\item{max.iter}{Maximum number of MIP iterations performed to update index
coefficients for a given model.}

\item{tol}{Tolerance for the objective function value (loss) of MIP.}

\item{tolCoefs}{Tolerance for coefficients.}

\item{TimeLimit}{A limit for the total time (in seconds) expended in a single
MIP iteration.}

\item{MIPGap}{Relative MIP optimality gap.}

\item{NonConvex}{The strategy for handling non-convex quadratic objectives or
non-convex quadratic constraints in Gurobi solver.}

\item{verbose}{The option to print detailed solver output.}

\item{parallel}{The option to use parallel processing in fitting SMI models
for different penalty parameter combinations.}

\item{workers}{If \code{parallel = TRUE}: Number of cores to use.}

\item{exclude.trunc}{The names of the predictor variables that should not be
truncated for stable predictions as a character string. (Since the
nonlinear functions are estimated using splines, extrapolation is not
desirable. Hence, if any predictor variable in \code{val.data} that is treated
non-linearly in the estimated model, will be truncated to be in the
in-sample range before obtaining predictions. If any variables are listed
here will be excluded from such truncation.)}

\item{recursive}{Whether to obtain recursive forecasts or not (default -
\code{FALSE}).}

\item{recursive_colRange}{If \code{recursive = TRUE}, the range of column
numbers in \code{val.data} to be filled with forecasts.
Recursive/autoregressive forecasting is required when the lags of the
response variable itself are used as predictor variables into the model.
Make sure such lagged variables are positioned together in increasing lag
order (i.e. \code{lag_1, lag_2, ..., lag_m}, \code{lag_m =} maximum lag
used) in \code{val.data}, with no break in the lagged variable sequence
even if some of the intermediate lags are not used as predictors.}
}
\value{
A list that contains six elements: \item{initial}{A list of
information of the model initialisation. (For descriptions of the list
elements see \code{\link{make_smimodelFit}}).} \item{best}{A list of
information of the final optimised model. (For descriptions of the list
elements see \code{\link{make_smimodelFit}}).} \item{best_lambdas}{Selected
penalty parameter combination.} \item{lambda0_seq}{Sequence of values for
lambda0 used to construct the initial grid.} \item{lambda2_seq}{Sequence of
values for lambda2 used to construct the initial grid.}
\item{searched}{A \code{tibble} containing the penalty parameter
combinations searched during the two-step greedy search and the
corresponding validation set MSEs.}
}
\description{
Function to perform a greedy search over a given grid of penalty parameter
combinations (lambda0, lambda2), and fits a single SMI model with the best
(lowest validation set MSE) penalty parameter combination. If the optimal
combination lies on the edge of the grid, the penalty parameters are adjusted
by ±10\%, and a second round of grid search is performed.This is a helper
function designed to be called from \code{\link{greedy_smimodel}}.
}
