% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_alpha.R
\name{init_alpha}
\alias{init_alpha}
\title{Initialising index coefficients}
\usage{
init_alpha(
  Y,
  X,
  index.ind,
  init.type = "penalisedReg",
  lambda0 = 1,
  lambda2 = 1,
  M = 10
)
}
\arguments{
\item{Y}{Column matrix of response.}

\item{X}{Matrix of predictors entering indices.}

\item{index.ind}{An integer vector that assigns group index for each
predictor.}

\item{init.type}{Type of initialisation for index coefficients.
("penalisedReg" - Penalised linear regression; "reg" - Linear regression)}

\item{lambda0}{Penalty parameter for L0 penalty.}

\item{lambda2}{Penalty parameter for L2 penalty.}

\item{M}{Big-M value used in MIP.}
}
\description{
Initialises index coefficient vector through linear regression or penalised
linear regression.
}
