% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_update.R
\name{inner_update}
\alias{inner_update}
\title{Updating index coefficients and non-linear functions iteratively}
\usage{
inner_update(
  x,
  data,
  yvar,
  index.vars,
  linear.vars,
  num_ind,
  dgz,
  alpha_old,
  lambda0 = 1,
  lambda2 = 1,
  M = 10,
  max.iter = 50,
  tol = 0.001,
  TimeLimit = Inf,
  verbose = FALSE
)
}
\arguments{
\item{x}{Fitted \code{gam}}

\item{data}{Training data set on which models will be trained. Should be a
\code{tibble}.}

\item{yvar}{Name of the response variable as a character string.}

\item{index.vars}{A character vector of names of the predictor variables for
which indices should be estimated.}

\item{linear.vars}{A character vector of names of the predictor variables
that should be included linearly into the model.}

\item{num_ind}{Number of indices.}

\item{dgz}{The \code{tibble} of derivatives of the estimated smooths.}

\item{alpha_old}{Current vector of index coefficients.}

\item{lambda0}{Penalty parameter for L0 penalty.}

\item{lambda2}{Penalty parameter for L2 penalty.}

\item{M}{Big-M value used in MIP.}

\item{max.iter}{Maximum number of MIP iterations performed to update index
coefficients for a given model.}

\item{tol}{Tolerance for loss.}

\item{TimeLimit}{A limit for the total time (in seconds) expended in a single
MIP iteration.}

\item{verbose}{The option to print detailed solver output.}
}
\description{
Iteratively updates index coefficients and non-linear functions using mixed
integer programming. (Used within \code{update_smimodel()}; users are not expected
to directly call this function.)
}
