% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_matrix.R
\name{lag_matrix}
\alias{lag_matrix}
\title{Function for adding lags of time series variables}
\usage{
lag_matrix(variable, n = 10)
}
\arguments{
\item{variable}{Variable to be lagged.}

\item{n}{Number of lags. The default value is \code{n = 10}.}
}
\value{
A \code{tibble}.
}
\description{
Generates specified number of lagged variables of the given variable in
the form of a tibble.
}
\examples{
# Adding lagged variables to an existing tibble
library(dplyr)
demand_data <- tsibbledata::vic_elec \%>\%
  mutate(Temperature_lag = lag_matrix(Temperature, 30)) \%>\%
  tidyr::unpack(Temperature_lag, names_sep = "_")
}
