% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{loss}
\alias{loss}
\title{Calculating the loss in MIP}
\usage{
loss(Y, Yhat, alpha, lambda0, lambda2)
}
\arguments{
\item{Y}{Column matrix of response.}

\item{Yhat}{Predicted value of the response.}

\item{alpha}{Vector of index coefficients.}

\item{lambda0}{Penalty parameter for L0 penalty.}

\item{lambda2}{Penalty parameter for L2 penalty.}
}
\description{
Calculates the value of the objective function (loss function) of the
corresponding mixed integer program.
}
