% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_smimodel.R
\name{make_smimodel}
\alias{make_smimodel}
\title{Converting a fitted \code{gam} object to a \code{smimodel} object}
\usage{
make_smimodel(
  x,
  yvar,
  index.vars,
  index.ind,
  index.data,
  index.names,
  alpha,
  linear.vars = NULL
)
}
\arguments{
\item{x}{A fitted \code{gam} object.}

\item{yvar}{Name of the response variable as a character string.}

\item{index.vars}{A character vector of names of the predictor variables for
which are estimated.}

\item{index.ind}{An integer vector that assigns group index for each
predictor in \code{index.vars}.}

\item{index.data}{A \code{tibble} including columns for the constructed indices.}

\item{index.names}{A character vector of names of the constructed indices.}

\item{alpha}{A vector of index coefficients.}

\item{linear.vars}{A character vector of names of the predictor variables
that are included linearly in the model.}
}
\description{
Converts a given object of class \code{gam} to an object of class \code{smimodel}.
}
