% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_gaim.R
\name{model_gaim}
\alias{model_gaim}
\title{Groupwise Additive Index Models (GAIM)}
\usage{
model_gaim(
  data,
  yvar,
  neighbour = 0,
  index.vars,
  index.ind,
  s.vars = NULL,
  linear.vars = NULL,
  ...
)
}
\arguments{
\item{data}{Training data set on which models will be trained. Must be a data
set of class \code{tsibble}.(Make sure there are no additional date or time
related variables except for the \code{index} of the \code{tsibble}). If
multiple models are fitted, the grouping variable should be the \code{key}
of the \code{tsibble}. If a \code{key} is not specified, a dummy key with
only one level will be created.}

\item{yvar}{Name of the response variable as a character string.}

\item{neighbour}{If multiple models are fitted: Number of neighbours of each
key (i.e. grouping variable) to be considered in model fitting to handle
smoothing over the key. Should be an \code{integer}. If \code{neighbour =
  x}, \code{x} number of keys before the key of interest and \code{x} number
of keys after the key of interest are grouped together for model fitting.
The default is \code{neighbour = 0} (i.e. no neighbours are considered for
model fitting).}

\item{index.vars}{A \code{character} vector of names of the predictor
variables for which indices should be estimated.}

\item{index.ind}{An \code{integer} vector that assigns group index for each
predictor in \code{index.vars}.}

\item{s.vars}{A \code{character} vector of names of the predictor variables
for which splines should be fitted individually (rather than considering as
part of an index).}

\item{linear.vars}{A \code{character} vector of names of the predictor
variables that should be included linearly into the model.}

\item{...}{Other arguments not currently used. (Note that the arguments in
\code{cgaim::cgaim()} related to constrained GAIMs are currently not
supported. Furthermore, the argument \code{subset} is also not supported
due to a bug in \code{cgaim::cgaim()}.)}
}
\value{
An object of class \code{gaimFit}. This is a \code{tibble} with two
columns: \item{key}{The level of the grouping variable (i.e. key of the
training data set).} \item{fit}{Information of the fitted model
corresponding to the \code{key}.} Each row of the column \code{fit} is an
object of class \code{cgaim}. For details refer \code{cgaim::cgaim()}.
}
\description{
A wrapper for \code{cgaim::cgaim()} enabling multiple GAIM models based on a
grouping variable. Currently does not support Constrained GAIM (CGAIM)s.
}
\details{
Group-wise Additive Index Model (GAIM) can be written in the form
\deqn{y_{i} = \sum_{j = 1}^{p} g_{j}(\boldsymbol{\alpha}_{j}^{T}\boldsymbol{x}_{ij}) +
\varepsilon_{i}, \quad i = 1, \dots, n,} where \eqn{y_{i}} is the univariate
response, \eqn{\boldsymbol{x}_{ij} \in \mathbb{R}^{l{j}}}, \eqn{j = 1,
  \dots, p} are pre-specified non-overlapping subsets of
\eqn{\boldsymbol{x}_{i}}, and \eqn{\boldsymbol{\alpha}_j} are the
corresponding index coefficients, \eqn{g_{j}} is an unknown (possibly
nonlinear) component function, and \eqn{\varepsilon_{i}} is the random
error, which is independent of \eqn{\boldsymbol{x}_{i}}.
}
\examples{
library(dplyr)
library(tibble)
library(tidyr)
library(tsibble)

# Simulate data
n = 1005
set.seed(123)
sim_data <- tibble(x_lag_000 = runif(n)) |>
  mutate(
    # Add x_lags
    x_lag = lag_matrix(x_lag_000, 5)) |>
  unpack(x_lag, names_sep = "_") |>
  mutate(
    # Response variable
    y = (0.9*x_lag_000 + 0.6*x_lag_001 + 0.45*x_lag_003)^3 + rnorm(n, sd = 0.1),
    # Add an index to the data set
    inddd = seq(1, n)) |>
  drop_na() |>
  select(inddd, y, starts_with("x_lag")) |>
  # Make the data set a `tsibble`
  as_tsibble(index = inddd)

# Predictors taken as index variables
index.vars <- colnames(sim_data)[3:7]

# Assign group indices for each predictor
index.ind = c(rep(1, 3), rep(2, 2))

# Predictors taken as non-linear variables not entering indices
s.vars = "x_lag_005"

# Model fitting
gaimModel <- model_gaim(data = sim_data,
                        yvar = "y",
                        index.vars = index.vars,
                        index.ind = index.ind,
                        s.vars = s.vars)
# Fitted model
gaimModel$fit[[1]]

}
\seealso{
\code{\link{model_smimodel}}, \code{\link{model_backward}},
\code{\link{model_ppr}}, \code{\link{model_gam}}, \code{\link{model_lm}}
}
