% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_gam.R
\name{model_gam}
\alias{model_gam}
\title{Generalised Additive Models}
\usage{
model_gam(
  data,
  yvar,
  family = gaussian(),
  neighbour = 0,
  s.vars,
  s.basedim = NULL,
  linear.vars = NULL,
  ...
)
}
\arguments{
\item{data}{Training data set on which models will be trained. Must be a data
set of class \code{tsibble}.(Make sure there are no additional
date/time/date-time/yearmonth/POSIXct/POSIXt variables except for the
\code{index} of the \code{tsibble}). If multiple models are fitted, the grouping
variable should be the key of the \code{tsibble}.}

\item{yvar}{Name of the response variable as a character string.}

\item{family}{A description of the error distribution and link function to be
used in the model (see \code{\link{glm}} and \code{\link{family}}).}

\item{neighbour}{If multiple models are fitted: Number of neighbours of each
key (i.e. grouping variable) to be considered in model fitting to handle
smoothing over the key. Should be an integer. If \code{neighbour = x}, \code{x}
number of keys before the key of interest and \code{x} number of keys after the
key of interest are grouped together for model fitting. The default is \code{0}
(i.e. no neighbours are considered for model fitting).}

\item{s.vars}{A character vector of names of the predictor variables for
which splines should be fitted (i.e. non-linear predictors).}

\item{s.basedim}{Dimension of the bases used to represent the smooth terms
corresponding to \code{s.vars}. (For more information refer \code{mgcv::s()}.)
(default: NULL)}

\item{linear.vars}{A character vector of names of the predictor variables
that should be included linearly into the model. (default:
NULL)}

\item{...}{Other arguments not currently used.}
}
\description{
A wrapper for \code{mgcv::gam()} enabling multiple GAMs based on a grouping
variable.
}
