% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_lm.R
\name{model_lm}
\alias{model_lm}
\title{Linear Regression models - This function is written only for the purpose
of testing Prediction Interval code}
\usage{
model_lm(data, yvar, neighbour = 0, linear.vars, ...)
}
\arguments{
\item{data}{Training data set on which models will be trained. Must be a data
set of class \code{tsibble}.(Make sure there are no additional
date/time/date-time/yearmonth/POSIXct/POSIXt variables except for the
\code{index} of the \code{tsibble}). If multiple models are fitted, the grouping
variable should be the key of the \code{tsibble}.}

\item{yvar}{Name of the response variable as a character string.}

\item{neighbour}{If multiple models are fitted: Number of neighbours of each
key (i.e. grouping variable) to be considered in model fitting to handle
smoothing over the key. Should be an integer. If \code{neighbour = x}, \code{x}
number of keys before the key of interest and \code{x} number of keys after the
key of interest are grouped together for model fitting. The default is \code{0}
(i.e. no neighbours are considered for model fitting).}

\item{linear.vars}{A character vector of names of the predictor variables
that should be included linearly into the model.}

\item{...}{Other arguments not currently used. (For more information on other
arguments that can be passed, refer \code{stats::ppr()}.)}
}
\description{
A wrapper for \code{stats::lm()} enabling multiple linear models based on a grouping
variable.
}
