% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_smimodel.R
\name{new_smimodel}
\alias{new_smimodel}
\title{Constructor function for the class \code{smimodel}}
\usage{
new_smimodel(
  data,
  yvar,
  index.vars,
  index.ind = NULL,
  index.coefs = NULL,
  linear.vars = NULL
)
}
\arguments{
\item{data}{Training data set on which models will be trained. Should be a
\code{tibble}.}

\item{yvar}{Name of the response variable as a character string.}

\item{index.vars}{A character vector of names of the predictor variables for
which indices should be estimated.}

\item{index.ind}{An integer vector that assigns group index for each
predictor in \code{index.vars}. (The default is \code{NULL}. If \code{NULL}, the model
will be initialised with an Additive Model.)}

\item{index.coefs}{A numeric vector of index coefficients (default: NULL).}

\item{linear.vars}{A character vector of names of the predictor variables
that should be included linearly into the model.}
}
\description{
Constructs an object of class \code{smimodel} using the information passed to
arguments.
}
