% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_smimodelFit.R
\name{new_smimodelFit}
\alias{new_smimodelFit}
\title{Constructor function for the class \code{smimodelFit}}
\usage{
new_smimodelFit(
  data,
  yvar,
  neighbour = 0,
  family = gaussian(),
  index.vars,
  initialise = c("additive", "linear", "userInput"),
  index.ind = NULL,
  index.coefs = NULL,
  s.vars = NULL,
  linear.vars = NULL
)
}
\arguments{
\item{data}{Training data set on which models will be trained. Should be a
\code{tsibble}.}

\item{yvar}{Name of the response variable as a character string.}

\item{neighbour}{If multiple models are fitted: Number of neighbours of each
key (i.e. grouping variable) to be considered in model fitting to handle
smoothing over the key. Should be an integer. If \code{neighbour = x}, \code{x}
number of keys before the key of interest and \code{x} number of keys after the
key of interest are grouped together for model fitting. The default is \code{0}
(i.e. no neighbours are considered for model fitting).}

\item{family}{A description of the error distribution and link function to be
used in the model (see \code{\link{glm}} and \code{\link{family}}).}

\item{index.vars}{A character vector of names of the predictor variables for
which indices should be estimated.}

\item{initialise}{The model structure with which the estimation process
should be initialised. The default is "additive", where the initial model
will be a nonparametric additive model. The other options are "linear" -
linear regression model (i.e. a special case single-index model, where the
initial values of the index coefficients are obtained through a linear
regression), and "userInput" - user specifies the initial model structure
(i.e. the number of indices and the placement of index variables among
indices) and the initial index coefficients through \code{index.ind} and
\code{index.coefs} arguments respectively.}

\item{index.ind}{If \code{initialise = "userInput"}: an integer vector that
assigns group index for each predictor in \code{index.vars}.}

\item{index.coefs}{If \code{initialise = "userInput"}: a numeric vector of index
coefficients.}

\item{s.vars}{A character vector of names of the predictor variables for
which splines should be fitted individually (rather than considering as a
part of an index considered in \code{index.vars}).}

\item{linear.vars}{A character vector of names of the predictor variables
that should be included linearly into the model.}
}
\description{
Constructs an object of class \code{smimodelFit} using the information passed to
arguments.
}
