% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_bootstrap.R
\name{pi_bootstrap}
\alias{pi_bootstrap}
\title{Prediction intervals using single season block bootstrapping}
\usage{
pi_bootstrap(
  object,
  newdata,
  season.period = 1,
  m = 1,
  num.futures = 1000,
  confidence = 95,
  recursive = FALSE,
  recursive_colRange = NULL
)
}
\arguments{
\item{object}{A fitted model object of the class \code{smimodel}, \code{backward},
\code{pprFit}, or \code{gaimFit}.}

\item{newdata}{The set of new data on for which the forecasts are required
(i.e. test set; should be a \code{tsibble}).}

\item{season.period}{Length of the seasonal period.}

\item{m}{Multiplier. (Block size = \code{season.period} * \code{m})}

\item{num.futures}{Number of possible future sample paths to be generated.}

\item{confidence}{Confidence level.}

\item{recursive}{Whether to obtain recursive forecasts or not (default -
FALSE).}

\item{recursive_colRange}{If \code{recursive = TRUE}, the range of column numbers
in \code{newdata} to be filled with forecasts.}
}
\description{
Generates multi-step prediction intervals corresponding to the forecasts
obtained on a given test set (i.e. \code{newdata}), using single season block
bootstrapping method.
}
\references{
Hyndman, R.J. & Fan, S. (2010). Density Forecasting for Long-Term
Peak Electricity Demand. \emph{IEEE Transactions on Power Systems}, 25(2),
1142–1153. \url{http://dx.doi.org/10.1109/TPWRS.2009.2036017}
}
