% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\docType{data}
\name{MAE}
\alias{MAE}
\alias{MSE}
\alias{point_measures}
\title{Point estimate accuracy measures}
\format{
An object of class \code{list} of length 2.
}
\usage{
MAE(residuals, na.rm = TRUE, ...)

MSE(residuals, na.rm = TRUE, ...)

point_measures
}
\arguments{
\item{residuals}{A vector of residuals from either the validation or test
data.}

\item{na.rm}{If \code{TRUE}, remove the missing values before calculating the
accuracy measure.}

\item{...}{Additional arguments for each measure.}
}
\value{
For the individual functions (\code{MAE}, \code{MSE}), returns a single numeric
scalar giving the requested accuracy measure.

For the exported object \code{point_measures}, returns a named list of functions
that can be supplied to higher-level accuracy routines.
}
\description{
Point estimate accuracy measures
}
\examples{
set.seed(123)
ytrain <- rnorm(100)
ytest  <- rnorm(30)
yhat   <- ytest + rnorm(30, sd = 0.3)
resid   <- ytest - yhat

MAE(resid)
MSE(resid)

}
\keyword{datasets}
