% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_intervals.R
\name{possibleFutures_benchmark}
\alias{possibleFutures_benchmark}
\title{Possible future sample paths (multi-step) from residuals of a fitted
benchmark model}
\usage{
possibleFutures_benchmark(object, newdata, bootstraps, recursive_colRange)
}
\arguments{
\item{object}{A fitted model object of the class \code{backward},
\code{pprFit}, or \code{gaimFit}.}

\item{newdata}{The set of new data on for which the forecasts are required
(i.e. test set; should be a \code{tsibble}).}

\item{bootstraps}{Generated matrix of bootstrapped residual series.}

\item{recursive_colRange}{The range of column numbers in \code{newdata} to be
filled with forecasts.}
}
\value{
A list containing the following components: \item{firstFuture}{A
\code{numeric} vector of 1-step-ahead simulated futures.}
\item{future_cols}{A list of multi-steps-ahead simulated futures, where
each list element corresponds to each 1-step-ahead simulated future in
\code{firstFuture}.}
}
\description{
Generates possible future sample paths (multi-step) using residuals of a
fitted benchmark model through recursive forecasting.
}
