% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.lmFit}
\alias{predict.lmFit}
\title{Obtaining forecasts on a test set from a fitted \code{lmFit}}
\usage{
\method{predict}{lmFit}(object, newdata, recursive = FALSE, recursive_colRange = NULL, ...)
}
\arguments{
\item{object}{A \code{lmFit} object.}

\item{newdata}{The set of new data on for which the forecasts are required
(i.e. test set; should be a \code{tsibble}).}

\item{recursive}{Whether to obtain recursive forecasts or not (default -
\code{FALSE}).}

\item{recursive_colRange}{If \code{recursive = TRUE}, the range of column
numbers in \code{newdata} to be filled with forecasts.
Recursive/autoregressive forecasting is required when the lags of the
response variable itself are used as predictor variables into the model.
Make sure such lagged variables are positioned together in increasing lag
order (i.e. \code{lag_1, lag_2, ..., lag_m}, \code{lag_m =} maximum lag
used) in \code{newdata}, with no break in the lagged variable sequence even
if some of the intermediate lags are not used as predictors.}

\item{...}{Other arguments not currently used.}
}
\value{
A \code{tsibble} with forecasts on test set.
}
\description{
Gives forecasts on a test set.
}
\examples{
library(dplyr)
library(tibble)
library(tidyr)
library(tsibble)

# Simulate data
n = 1015
set.seed(123)
sim_data <- tibble(x_lag_000 = runif(n)) |>
  mutate(
    # Add x_lags
    x_lag = lag_matrix(x_lag_000, 5)) |>
  unpack(x_lag, names_sep = "_") |>
  mutate(
    # Response variable
    y = (0.9*x_lag_000 + 0.6*x_lag_001 + 0.45*x_lag_003)^3 + rnorm(n, sd = 0.1),
    # Add an index to the data set
    inddd = seq(1, n)) |>
  drop_na() |>
  select(inddd, y, starts_with("x_lag")) |>
  # Make the data set a `tsibble`
  as_tsibble(index = inddd)
  
# Training set
sim_train <- sim_data[1:1000, ]
# Test set
sim_test <- sim_data[1001:1010, ]

# Predictor variables
linear.vars <- colnames(sim_data)[3:8]

# Model fitting
lmModel <- model_lm(data = sim_train,
                    yvar = "y",
                    linear.vars = linear.vars)
                    
predict(object = lmModel, newdata = sim_test)

}
