% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.smimodel.R
\name{predict.smimodel}
\alias{predict.smimodel}
\title{Obtaining forecasts on a test set from a fitted \code{smimodel}}
\usage{
\method{predict}{smimodel}(
  object,
  newdata,
  data,
  recursive = FALSE,
  recursive_colRange = NULL,
  ...
)
}
\arguments{
\item{object}{A \code{smimodel} object.}

\item{newdata}{The set of new data on for which the forecasts are required
(i.e. test set; should be a \code{tibble}).}

\item{data}{Training data set on which models will be trained. Should be a
\code{tibble}.}

\item{recursive}{Whether to obtain recursive forecasts or not (default -
FALSE).}

\item{recursive_colRange}{If \code{recursive = TRUE}, The range of column numbers
in \code{newdata} to be filled with forecasts.}

\item{...}{Other arguments not currently used.}
}
\description{
Gives forecasts on a test set.
}
