% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.smimodelFit}
\alias{predict.smimodelFit}
\title{Obtaining forecasts on a test set from a \code{smimodelFit}}
\usage{
\method{predict}{smimodelFit}(
  object,
  newdata,
  exclude.trunc = NULL,
  recursive = FALSE,
  recursive_colRange = NULL,
  ...
)
}
\arguments{
\item{object}{A \code{smimodelFit} object.}

\item{newdata}{The set of new data on for which the forecasts are required
(i.e. test set; should be a \code{tsibble}).}

\item{exclude.trunc}{The names of the predictor variables that should not be
truncated for stable predictions as a character string. (Since the
nonlinear functions are estimated using splines, extrapolation is not
desirable. Hence, if any predictor variable in \code{newdata} that is treated
non-linearly in the estimated model, will be truncated to be in the
in-sample range before obtaining predictions. If any variables are listed
here will be excluded from such truncation.)}

\item{recursive}{Whether to obtain recursive forecasts or not (default -
\code{FALSE}).}

\item{recursive_colRange}{If \code{recursive = TRUE}, the range of column
numbers in \code{newdata} to be filled with forecasts.
Recursive/autoregressive forecasting is required when the lags of the
response variable itself are used as predictor variables into the model.
Make sure such lagged variables are positioned together in increasing lag
order (i.e. \code{lag_1, lag_2, ..., lag_m}, \code{lag_m =} maximum lag
used) in \code{newdata}, with no break in the lagged variable sequence even
if some of the intermediate lags are not used as predictors.}

\item{...}{Other arguments not currently used.}
}
\value{
A \code{tibble} with forecasts on test set.
}
\description{
Gives forecasts on a test set.
}
