% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_intervals.R
\name{prep_newdata}
\alias{prep_newdata}
\title{Prepare a data set for recursive forecasting}
\usage{
prep_newdata(newdata, recursive_colRange)
}
\arguments{
\item{newdata}{Data set to be ared. Should be a \code{tsibble}.}

\item{recursive_colRange}{The range of column numbers (lagged response
columns) in \code{newdata} from which existing values should be removed.
Make sure such columns are positioned together in increasing lag order
(i.e. \code{lag_1, lag_2, ..., lag_m}, \code{lag_m =} maximum lag used) in
\code{newdata}, with no break in the lagged variable sequence even if some
of the intermediate lags are not used as predictors.}
}
\value{
A \code{tibble}.
}
\description{
Prepare a test data for recursive forecasting by appropriately removing
existing (actual) values from a specified range of columns (lagged response
columns) of the data set. Handles seasonal data with gaps.
}
