% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.backward}
\alias{print.backward}
\title{Printing a \code{backward} object}
\usage{
\method{print}{backward}(x, ...)
}
\arguments{
\item{x}{A model object of class \code{backward} as produced by
\code{model_backward()} function.}

\item{...}{Other arguments not currently used.}
}
\value{
No return value; called for side effects. Prints a summary of the
fitted model(s) to console.
}
\description{
The default print method for a \code{backward} object.
}
