% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.groupSmimodel.R
\name{print.groupSmimodel}
\alias{print.groupSmimodel}
\title{Printing a \code{groupSmimodel} object}
\usage{
\method{print}{groupSmimodel}(x, ...)
}
\arguments{
\item{x}{A model object of class \code{groupSmimodel} as produced by
\code{new_groupSmimodel()} or \code{update_groupSmimodel()} functions.}

\item{...}{Other arguments not currently used.}
}
\description{
The default print method for a \code{groupSmimodel} object.
}
