% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.pprFit}
\alias{print.pprFit}
\title{Printing a \code{pprFit} object}
\usage{
\method{print}{pprFit}(x, ...)
}
\arguments{
\item{x}{A model object of class \code{pprFit} as produced by \code{model_ppr()}
function.}

\item{...}{Other arguments not currently used.}
}
\description{
The default print method for a \code{pprFit} object.
}
