% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.smimodelFit}
\alias{print.smimodelFit}
\title{Printing a \code{smimodelFit} object}
\usage{
\method{print}{smimodelFit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{smimodelFit}.}

\item{...}{Other arguments not currently used.}
}
\description{
The default print method for a \code{smimodelFit} object.
}
