% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_intervals.R
\name{randomBlock}
\alias{randomBlock}
\title{Randomly sampling a block}
\usage{
randomBlock(series, block.size)
}
\arguments{
\item{series}{A series from which a block should be sampled.}

\item{block.size}{Size of the block to be sampled.}
}
\description{
Samples a block of specified size from a given series starting form a random
point in the series.
}
