% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_intervals.R
\name{remove_lags}
\alias{remove_lags}
\title{Remove actual values from a data set for recursive forecasting}
\usage{
remove_lags(data, recursive_colRange)
}
\arguments{
\item{data}{Data set (a \code{tibble}) from which the actual lagged values
should be removed.}

\item{recursive_colRange}{The range of column numbers in \code{data} from
which lagged values should be removed.}
}
\value{
A \code{tibble}.
}
\description{
Appropriately removes exisiting (actual) values from the specified column
range (lagged response columns) of a given data set (typicall a test set for
which recursive forecasting is required).
}
