% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_intervals.R
\name{residBootstrap}
\alias{residBootstrap}
\title{Generate multiple single season block bootstrap series}
\usage{
residBootstrap(x, season.period = 1, m = 1, num.bootstrap = 1000)
}
\arguments{
\item{x}{A series of residuals from which bootstrap series to be generated.}

\item{season.period}{Length of the seasonal period.}

\item{m}{Multiplier. (Block size = \code{season.period} * \code{m})}

\item{num.bootstrap}{Number of bootstrap series to be generated.}
}
\description{
Generates multiple replications of single season block bootstrap series.
}
