% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.smimodel}
\alias{residuals.smimodel}
\title{Extract residuals from a fitted \code{smimodel}}
\usage{
\method{residuals}{smimodel}(object, ...)
}
\arguments{
\item{object}{A \code{smimodel} object.}

\item{...}{Other arguments not currently used.}
}
\description{
Generates residuals from a fitted \code{smimodel} object.
}
