% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_smimodel.R
\name{scaling}
\alias{scaling}
\title{Scale data}
\usage{
scaling(data, index.vars)
}
\arguments{
\item{data}{Training data set on which models will be trained. Should be a
\code{tibble}.}

\item{index.vars}{A character vector of names of the predictor variables for
which indices should be estimated.}
}
\value{
A list containing the following components: \item{scaled_data}{The
scaled data set of class \code{tibble}.} \item{scaled_info}{A named
\code{numeric} vector of standard deviations of \code{index.vars} that were
used to scale the corresponding columns of \code{data}.}
}
\description{
Scales the columns of the \code{data} corresponding to \code{index.vars}.
}
