% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_bootstrap.R
\name{seasonBootstrap}
\alias{seasonBootstrap}
\title{Single season block bootstrap}
\usage{
seasonBootstrap(x, season.period = 1, m = 1)
}
\arguments{
\item{x}{A series of residuals from which bootstrap series to be generated.}

\item{season.period}{Length of the seasonal period.}

\item{m}{Multiplier. (Block size = \code{season.period} * \code{m})}
}
\description{
Generates a single replication of single season block bootstrap series.
}
