% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smimodel-package.R
\docType{package}
\name{smimodel-package}
\alias{smimodel}
\alias{smimodel-package}
\title{smimodel: Sparse Multiple Index Models for Nonparametric Forecasting}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Implements a general algorithm for estimating Sparse Multiple Index (SMI) models for nonparametric forecasting and prediction. Estimation of SMI models requires the Gurobi mixed integer programming (MIP) solver via the gurobi R package. To use this functionality, the Gurobi Optimizer must be installed, and a valid license obtained and activated from \url{https://www.gurobi.com}. The gurobi R package must then be installed and configured following the instructions at \url{https://support.gurobi.com/hc/en-us/articles/14462206790033-How-do-I-install-Gurobi-for-R}. The package also includes functions for fitting nonparametric additive models with backward elimination, group-wise additive index models, and projection pursuit regression models as benchmark comparison methods. In addition, it provides tools for generating prediction intervals to quantify uncertainty in point forecasts produced by the SMI model and benchmark models, using the classical block bootstrap and a new method called conformal bootstrap, which integrates block bootstrap with split conformal prediction.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nuwani-palihawadana/smimodel}
  \item \url{https://nuwani-palihawadana.github.io/smimodel/}
  \item Report bugs at \url{https://github.com/nuwani-palihawadana/smimodel/issues}
}

}
\author{
\strong{Maintainer}: Nuwani Palihawadana \email{nuwanipalihawadana@gmail.com} (\href{https://orcid.org/0009-0008-6395-7797}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Xiaoqian Wang \email{Xiaoqian.Wang@amss.ac.cn} (\href{https://orcid.org/0000-0003-4827-496X}{ORCID}) [contributor]
}

}
\keyword{package}
