% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smimodel.R
\name{smimodel}
\alias{smimodel}
\title{Sparse Multiple Index (SMI) models}
\usage{
smimodel(
  data,
  yvar,
  index.vars,
  index.ind = NULL,
  index.coefs = NULL,
  linear.vars = NULL,
  lambda0 = 1,
  lambda2 = 1,
  M = 10,
  max.iter = 50,
  tol = 0.001,
  tolCoefs = 0.001,
  TimeLimit = Inf,
  verbose = FALSE
)
}
\arguments{
\item{data}{Training data set on which models will be trained. Should be a
\code{tibble}.}

\item{yvar}{Name of the response variable as a character string.}

\item{index.vars}{A character vector of names of the predictor variables for
which indices should be estimated.}

\item{index.ind}{An integer vector that assigns group index for each
predictor in \code{index.vars}. (The default is \code{NULL}. If \code{NULL}, the model
will be initialised with an Additive Model.)}

\item{index.coefs}{A numeric vector of index coefficients (default: NULL).}

\item{linear.vars}{A character vector of names of the predictor variables
that should be included linearly into the model.}

\item{lambda0}{Penalty parameter for L0 penalty.}

\item{lambda2}{Penalty parameter for L2 penalty.}

\item{M}{Big-M value used in MIP.}

\item{max.iter}{Maximum number of MIP iterations performed to update index
coefficients for a given model.}

\item{tol}{Tolerance for loss.}

\item{tolCoefs}{Tolerance for coefficients.}

\item{TimeLimit}{A limit for the total time (in seconds) expended in a single
MIP iteration.}

\item{verbose}{The option to print detailed solver output.}
}
\description{
Fits a nonparametric multiple index model to data with simultaneous variable
selection (hence "sparse").
}
