% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_index.R
\name{split_index}
\alias{split_index}
\title{Splitting predictors into multiple indices}
\usage{
split_index(num_pred, num_ind)
}
\arguments{
\item{num_pred}{Number of predictors.}

\item{num_ind}{Number of indices.}
}
\description{
Splits a given number of predictors into a given number of indices.
}
