% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_alpha.R
\name{update_alpha}
\alias{update_alpha}
\title{Updating index coefficients using MIP}
\usage{
update_alpha(
  Y,
  X,
  num_pred,
  num_ind,
  index.ind,
  dgz,
  alpha_old,
  lambda0 = 1,
  lambda2 = 1,
  M = 10,
  TimeLimit = Inf,
  MIPGap = 1e-04,
  NonConvex = -1,
  verbose = FALSE
)
}
\arguments{
\item{Y}{Column matrix of response.}

\item{X}{Matrix of predictors (size adjusted to number of indices).}

\item{num_pred}{Number of predictors.}

\item{num_ind}{Number of indices.}

\item{index.ind}{An integer vector that assigns group index for each
predictor.}

\item{dgz}{The \code{tibble} of derivatives of the estimated smooths from previous
iteration.}

\item{alpha_old}{Vector of index coefficients from previous iteration.}

\item{lambda0}{Penalty parameter for L0 penalty.}

\item{lambda2}{Penalty parameter for L2 penalty.}

\item{M}{Big-M value used in MIP.}

\item{TimeLimit}{A limit for the total time (in seconds) expended in a single
MIP iteration.}

\item{MIPGap}{Relative MIP optimality gap.}

\item{NonConvex}{The strategy for handling non-convex quadratic objectives or
non-convex quadratic constraints in Gurobi solver.}

\item{verbose}{The option to print detailed solver output.}
}
\description{
Updates index coefficients by solving a mixed integer program.
}
