% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_groupSmimodel.R
\name{update_groupSmimodel}
\alias{update_groupSmimodel}
\title{Updating a \code{groupSmimodel}}
\usage{
update_groupSmimodel(
  object,
  data,
  neighbour = 0,
  lambda0 = 1,
  lambda2 = 1,
  M = 10,
  max.iter = 50,
  tol = 0.001,
  TimeLimit = Inf,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{object}{A \code{groupSmimodel} object}

\item{data}{Training data set on which models will be trained. Must be a
data set of class \code{tsibble}.(Make sure there are no additional
date/time/date-time/yearmonth/POSIXct/POSIXt variables except for the
\code{index} of the \code{tsibble}). If multiple models are fitted, the grouping
variable should be the key of the \code{tsibble}.}

\item{neighbour}{If multiple models are fitted: Number of neighbours of each
key (i.e. grouping variable) to be considered in model fitting to handle
smoothing over the key. Should be an integer. If \code{neighbour = x}, \code{x}
number of keys before the key of interest and \code{x} number of keys after the
key of interest are grouped together for model fitting. The default is \code{0}
(i.e. no neighbours are considered for model fitting).}

\item{lambda0}{Penalty parameter for L0 penalty.}

\item{lambda2}{Penalty parameter for L2 penalty.}

\item{M}{Big-M value used in MIP.}

\item{max.iter}{Maximum number of MIP iterations performed to update index
coefficients for a given model.}

\item{tol}{Tolerance for loss.}

\item{TimeLimit}{A limit for the total time (in seconds) expended in a single
MIP iteration.}

\item{verbose}{The option to print detailed solver output.}

\item{...}{Other arguments not currently used.}
}
\description{
Optimises and updates a given \code{groupSmimodel}.
}
