% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_smimodel.R
\name{update_smimodel}
\alias{update_smimodel}
\title{Updating a \code{smimodel}}
\usage{
update_smimodel(
  object,
  data,
  lambda0 = 1,
  lambda2 = 1,
  M = 10,
  max.iter = 50,
  tol = 0.001,
  tolCoefs = 0.001,
  TimeLimit = Inf,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{object}{A \code{smimodel} object.}

\item{data}{Training data set on which models will be trained. Should be a
\code{tibble}.}

\item{lambda0}{Penalty parameter for L0 penalty.}

\item{lambda2}{Penalty parameter for L2 penalty.}

\item{M}{Big-M value used in MIP.}

\item{max.iter}{Maximum number of MIP iterations performed to update index
coefficients for a given model.}

\item{tol}{Tolerance for loss.}

\item{tolCoefs}{Tolerance for coefficients.}

\item{TimeLimit}{A limit for the total time (in seconds) expended in a single
MIP iteration.}

\item{verbose}{The option to print detailed solver output.}

\item{...}{Other arguments not currently used.}
}
\description{
Optimises and updates a given \code{smimodel}.
}
