% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{augment.backward}
\alias{augment.backward}
\title{Augment function for class \code{backward}}
\usage{
\method{augment}{backward}(x, ...)
}
\arguments{
\item{x}{A \code{backward} object.}

\item{...}{Other arguments not currently used.}
}
\value{
A \code{tibble}.
}
\description{
Generates residuals and fitted values of a fitted \code{backward} object.
}
