% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{augment.gaimFit}
\alias{augment.gaimFit}
\title{Augment function for class \code{gaimFit}}
\usage{
\method{augment}{gaimFit}(x, ...)
}
\arguments{
\item{x}{A \code{gaimFit} object.}

\item{...}{Other arguments not currently used.}
}
\description{
Generates residuals and fitted values of a fitted \code{gaimFit} object.
}
