% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{augment.pprFit}
\alias{augment.pprFit}
\title{Augment function for class \code{pprFit}}
\usage{
\method{augment}{pprFit}(x, ...)
}
\arguments{
\item{x}{A \code{pprFit} object.}

\item{...}{Other arguments not currently used.}
}
\value{
A \code{tibble}.
}
\description{
Generates residuals and fitted values of a fitted \code{pprFit} object.
}
\examples{
library(dplyr)
library(tibble)
library(tidyr)
library(tsibble)

# Simulate data
n = 1005
set.seed(123)
sim_data <- tibble(x_lag_000 = runif(n)) |>
  mutate(
    # Add x_lags
    x_lag = lag_matrix(x_lag_000, 5)) |>
  unpack(x_lag, names_sep = "_") |>
  mutate(
    # Response variable
    y = (0.9*x_lag_000 + 0.6*x_lag_001 + 0.45*x_lag_003)^3 + rnorm(n, sd = 0.1),
    # Add an index to the data set
    inddd = seq(1, n)) |>
  drop_na() |>
  select(inddd, y, starts_with("x_lag")) |>
  # Make the data set a `tsibble`
  as_tsibble(index = inddd)

# Index variables
index.vars <- colnames(sim_data)[3:8]

# Model fitting
pprModel <- model_ppr(data = sim_data,
                      yvar = "y",
                      index.vars = index.vars)

# Obtain residuals and fitted values
augment(pprModel)

}
