% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{augment.smimodel}
\alias{augment.smimodel}
\title{Augment function for class \code{smimodel}}
\usage{
\method{augment}{smimodel}(x, ...)
}
\arguments{
\item{x}{A \code{smimodel} object.}

\item{...}{Other arguments not currently used.}
}
\value{
A \code{tibble}.
}
\description{
Generates residuals and fitted values of a fitted \code{smimodel} object.
}
