% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.smimodel}
\alias{autoplot.smimodel}
\title{Plot estimated smooths from a fitted \code{smimodel}}
\usage{
\method{autoplot}{smimodel}(object, model = 1, ...)
}
\arguments{
\item{object}{A \code{smimodel} object returned from \code{model_smimodel()}.}

\item{model}{An \code{integer} to indicate the smooths of which model (out of the
set of multiple models fitted) to be plotted.}

\item{...}{Other arguments not currently used.}
}
\value{
Plot(s) of fitted spline(s).
}
\description{
Plots the graphs of fitted spline(s) for a specified predictor variable, by
fixing all the other predictor values. If a set of multiple models are
fitted, plots individual graphs of fitted values of all the models in a
single plot except for tensor smooths fitted in \code{model_distlag()}. out of the
set of multiple models fitted)
}
