% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.smimodel}
\alias{autoplot.smimodel}
\title{Plot estimated smooths from a fitted \code{smimodel}}
\usage{
\method{autoplot}{smimodel}(object, model = 1, ...)
}
\arguments{
\item{object}{A \code{smimodel} object.}

\item{model}{An \code{integer} to indicate the smooths of which model (out of
the set of multiple models fitted) to be plotted.}

\item{...}{Other arguments not currently used.}
}
\value{
Plot(s) of fitted spline(s).
}
\description{
Plots the graphs of fitted spline(s). If a set of multiple models are fitted,
plots graphs of fitted spline(s) of a specified model (in argument
\code{model}) out of the set of multiple models fitted.
}
\examples{
\dontrun{
library(dplyr)
library(ROI)
library(tibble)
library(tidyr)
library(tsibble)

# Simulate data
n = 1005
set.seed(123)
sim_data <- tibble(x_lag_000 = runif(n)) |>
  mutate(
    # Add x_lags
    x_lag = lag_matrix(x_lag_000, 5)) |>
  unpack(x_lag, names_sep = "_") |>
  mutate(
    # Response variable
    y = (0.9*x_lag_000 + 0.6*x_lag_001 + 0.45*x_lag_003)^3 + rnorm(n, sd = 0.1),
    # Add an index to the data set
    inddd = seq(1, n)) |>
  drop_na() |>
  select(inddd, y, starts_with("x_lag")) |>
  # Make the data set a `tsibble`
  as_tsibble(index = inddd)

# Index variables
index.vars <- colnames(sim_data)[3:8]

# Model fitting
smimodel_ppr <- model_smimodel(data = sim_data,
                               yvar = "y",
                               index.vars = index.vars,
                               initialise = "ppr")

autoplot(smimodel_ppr)
}

}
