% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb.R
\name{blockBootstrap}
\alias{blockBootstrap}
\title{Futures through single season block bootstrapping}
\usage{
blockBootstrap(
  object,
  newdata,
  resids,
  preds,
  season.period = 1,
  m = 1,
  num.futures = 1000,
  recursive = FALSE,
  recursive_colRange = NULL
)
}
\arguments{
\item{object}{Fitted model object.}

\item{newdata}{Test data set. Must be a data set of class \code{tsibble}.}

\item{resids}{In-sample residuals from the fitted model.}

\item{preds}{Predictions for the test set (i.e. data for the forecast
horizon).}

\item{season.period}{Length of the seasonal period.}

\item{m}{Multiplier. (Block size = \code{season.period} * \code{m})}

\item{num.futures}{Number of possible future sample paths to be generated.}

\item{recursive}{Whether to obtain recursive forecasts or not (default -
FALSE).}

\item{recursive_colRange}{If \code{recursive = TRUE}, The range of column numbers
in test data to be filled with forecasts.}
}
\description{
Gerenates possible future sample paths by applying the single season block
bootstrap method.
}
