% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_bootstrap.R
\name{coverage}
\alias{coverage}
\title{Coverage of a calculated prediction interval}
\usage{
coverage(bounds, newdata, yvar)
}
\arguments{
\item{bounds}{The list returned from \code{pi_bootstrap()}.}

\item{newdata}{The set of new data on for which the forecasts are required
(i.e. test set; should be a \code{tsibble}).}

\item{yvar}{Response variable as a character string.}
}
\description{
Computes the actual coverage probability of a calculated prediction interval.
}
