% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eliminate.R
\name{eliminate}
\alias{eliminate}
\title{Function to eliminate a specified variable and fit a nonparametric additive
model with remaining variables}
\usage{
eliminate(
  ind,
  train,
  val,
  yvar,
  family = gaussian(),
  log.transformed = FALSE,
  s.vars = NULL,
  s.basedim = NULL,
  linear.vars = NULL
)
}
\arguments{
\item{ind}{An integer corresponding to the position of the predictor variable
to be eliminated when fitting the model. (i.e. the function will combine
\code{s.vars} and \code{linear.vars} in a single vector and eliminate the element
corresponding to \code{ind}.)}

\item{train}{The data set on which the model(s) will be trained. Must be a
data set of class \code{tsibble}.}

\item{val}{Validation data set. (The data set on which the model selection
will be performed.) Must be a data set of class \code{tsibble}.}

\item{yvar}{Name of the response variable as a character string.}

\item{family}{A description of the error distribution and link function to be
used in the model (see \code{glm} and \code{family}).}

\item{log.transformed}{Whether the response is log-transformed or not.}

\item{s.vars}{A character vector of names of the predictor variables for
which splines should be fitted (i.e. non-linear predictors). (default:
NULL)}

\item{s.basedim}{Dimension of the bases used to represent the smooth terms
corresponding to \code{s.vars}. (For more information refer \code{mgcv::s()}.)
(default: NULL)}

\item{linear.vars}{A character vector of names of the predictor variables
that should be included linearly into the model (i.e. linear predictors).
(default: NULL)}
}
\description{
This is an internal function of the package \code{smimodel}, and designed to be
called from \code{backward()}.
}
