% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast.backward}
\alias{forecast.backward}
\title{Forecasting using nonparametric additive models with backward elimination}
\usage{
\method{forecast}{backward}(
  object,
  h = 1,
  level = c(80, 95),
  newdata,
  exclude.trunc = NULL,
  recursive = FALSE,
  recursive_colRange = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{backward}. Usually the result of a
call to \code{\link{model_backward}}.}

\item{h}{Forecast horizon.}

\item{level}{Confidence level for prediction intervals.}

\item{newdata}{The set of new data on for which the forecasts are required
(i.e. test set; should be a \code{tsibble}).}

\item{exclude.trunc}{The names of the predictor variables that should not be
truncated for stable predictions as a character string.}

\item{recursive}{Whether to obtain recursive forecasts or not (default -
\code{FALSE}).}

\item{recursive_colRange}{If \code{recursive = TRUE}, the range of column
numbers in \code{newdata} to be filled with forecasts.}

\item{...}{Other arguments not currently used.}
}
\value{
An object of class \code{forecast}. Here, it is a list containing the
following elements: \item{method}{The name of the forecasting method as a
character string.} \item{model}{The fitted model.} \item{mean}{Point
forecasts as a time series.} \item{residuals}{Residuals from the fitted
model.} \item{fitted}{Fitted values (one-step forecasts).}
}
\description{
Returns forecasts and other information for nonparametric additive models
with backward elimination.
}
