% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy.fit.R
\name{greedy.fit}
\alias{greedy.fit}
\title{Greedy search for tuning penalty parameters}
\usage{
greedy.fit(
  data,
  yvar,
  neighbour = 0,
  family = gaussian(),
  index.vars,
  initialise = c("ppr", "additive", "linear", "multiple", "userInput"),
  num_ind = 5,
  num_models = 5,
  seed = 123,
  index.ind = NULL,
  index.coefs = NULL,
  s.vars = NULL,
  linear.vars = NULL,
  lambda0_seq,
  lambda2_seq,
  lambda_step,
  lambda0_start_seq,
  lambda2_start_seq,
  M = 10,
  max.iter = 50,
  tol = 0.001,
  tolCoefs = 0.001,
  TimeLimit = Inf,
  MIPGap = 1e-04,
  NonConvex = -1,
  verbose = FALSE,
  parallel = FALSE,
  workers = NULL
)
}
\arguments{
\item{data}{Training data set on which models will be trained. Should be a
\code{tibble}.}

\item{yvar}{Name of the response variable as a character string.}

\item{neighbour}{If multiple models are fitted: Number of neighbours of each
key (i.e. grouping variable) to be considered in model fitting to handle
smoothing over the key. Should be an integer. If \code{neighbour = x}, \code{x}
number of keys before the key of interest and \code{x} number of keys after the
key of interest are grouped together for model fitting. The default is \code{0}
(i.e. no neighbours are considered for model fitting).}

\item{family}{A description of the error distribution and link function to be
used in the model (see \code{\link{glm}} and \code{\link{family}}).}

\item{index.vars}{A character vector of names of the predictor variables for
which indices should be estimated.}

\item{initialise}{The model structure with which the estimation process
should be initialised. The default is "ppr", where the initial model is
derived from projection pursuit regression. The other options are
"additive" - nonparametric additive model, "linear" - linear regression
model (i.e. a special case single-index model, where the initial values of
the index coefficients are obtained through a linear regression),
"multiple" - multiple models are fitted starting with different initial
models (default \code{num_ind} (number of indices) = 5; five random instances of
the model (i.e. the predictor assignment to indices and initial index
coefficients are generated randomly) are considered), and the final optimal
model with the lowest loss is returned (user can change the number of
indices considered using \code{num_ind} argument), and "userInput" - user
specifies the initial model structure (i.e. the number of indices and the
placement of index variables among indices) and the initial index
coefficients through \code{index.ind} and \code{index.coefs} arguments respectively.}

\item{num_ind}{If \verb{initialise = "ppr" or "multiple"}: an integer that
specifies the number of indices to be used in the model(s).}

\item{num_models}{If \code{initialise = "multiple"}: an integer that specifies the
number of starting points to be checked.}

\item{seed}{If \code{initialise = "multiple"}: the seed to be set when generating
random starting points.}

\item{index.ind}{If \code{initialise = "userInput"}: an integer vector that
assigns group index for each predictor in \code{index.vars}.}

\item{index.coefs}{If \code{initialise = "userInput"}: a numeric vector of index
coefficients.}

\item{s.vars}{A character vector of names of the predictor variables for
which splines should be fitted individually (rather than considering as a
part of an index considered in \code{index.vars}).}

\item{linear.vars}{A character vector of names of the predictor variables
that should be included linearly into the model.}

\item{lambda0_seq}{A numeric vector of candidate values for lambda0 (penalty
parameter for L0 penalty).}

\item{lambda2_seq}{A numeric vector of candidate values for lambda2 (penalty
parameter for L2 penalty).}

\item{lambda_step}{Step size between two adjacent values in \code{lambda0_seq} and
\code{lambda2_seq}.}

\item{lambda0_start_seq}{A subset from \code{lambda0_seq} as candidate starting
points for the greedy search.}

\item{lambda2_start_seq}{A subset from \code{lambda2_seq} as candidate starting
points for the greedy search.}

\item{M}{Big-M value used in MIP.}

\item{max.iter}{Maximum number of MIP iterations performed to update index
coefficients for a given model.}

\item{tol}{Tolerance for loss.}

\item{tolCoefs}{Tolerance for coefficients.}

\item{TimeLimit}{A limit for the total time (in seconds) expended in a single
MIP iteration.}

\item{MIPGap}{Relative MIP optimality gap.}

\item{NonConvex}{The strategy for handling non-convex quadratic objectives or
non-convex quadratic constraints in Gurobi solver.}

\item{verbose}{The option to print detailed solver output.}

\item{parallel}{The option to use parallel processing in fitting \code{smimodel}s
for different penalty parameter combinations.}

\item{workers}{If \code{parallel = TRUE}: Number of cores to use.}
}
\description{
Function to perform a greedy search over a given grid of penalty parameter
combinations (lambda0, lambda2), and fits a single SMI model with the best
(lowest in-sample MSE) penalty parameter combination. This is a helper
function designed to be called from \code{greedy_smimodel()}.
}
