% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_smimodel.R
\name{init_alpha}
\alias{init_alpha}
\title{Initialising index coefficients}
\usage{
init_alpha(
  Y,
  X,
  index.ind,
  init.type = "penalisedReg",
  lambda0 = 1,
  lambda2 = 1,
  M = 10
)
}
\arguments{
\item{Y}{Column matrix of response.}

\item{X}{Matrix of predictors entering indices.}

\item{index.ind}{An \code{integer} vector that assigns group index for each
predictor.}

\item{init.type}{Type of initialisation for index coefficients.
(\code{"penalisedReg"} - Penalised linear regression; \code{"reg"} - Linear
regression)}

\item{lambda0}{If \code{init.type = "penalisedReg"}, penalty parameter for L0
penalty.}

\item{lambda2}{If \code{init.type = "penalisedReg"}, penalty parameter for L2
penalty.}

\item{M}{If \code{init.type = "penalisedReg"}, the big-M value to be used in
the MIP.}
}
\value{
A list containing the following components:
\item{alpha_init}{Normalised vector of index coefficients.}
\item{alpha_nonNormalised}{Non-normalised (i.e. prior to normalising)
vector of index coefficients.}
}
\description{
Initialises index coefficient vector through linear regression or penalised
linear regression.
}
