% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_matrix.R
\name{lag_matrix}
\alias{lag_matrix}
\title{Function for adding lags of time series variables}
\usage{
lag_matrix(variable, n = 10)
}
\arguments{
\item{variable}{Variable to be lagged.}

\item{n}{Number of lags. The default value is \code{n = 10}.}
}
\value{
A \code{tibble}.
}
\description{
Generates specified number of lagged variables of the given variable in the
form of a tibble.
}
\examples{
library(dplyr)
library(tibble)
library(tidyr)
# Adding lagged variables to an existing tibble
set.seed(123)
sim_data <- tibble(x_lag_000 = runif(100)) |>
  mutate(x_lag = lag_matrix(x_lag_000, 3)) |>
  unpack(x_lag, names_sep = "_")
}
