% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_smimodelFit.R
\name{make_smimodelFit}
\alias{make_smimodelFit}
\title{Converting a fitted \code{gam} object to a \code{smimodelFit} object}
\usage{
make_smimodelFit(
  x,
  yvar,
  neighbour,
  index.vars,
  index.ind,
  index.data,
  index.names,
  alpha,
  s.vars = NULL,
  linear.vars = NULL
)
}
\arguments{
\item{x}{A fitted \code{gam} object.}

\item{yvar}{Name of the response variable as a character string.}

\item{neighbour}{If multiple models are fitted: Number of neighbours of each
key (i.e. grouping variable) to be considered in model fitting to handle
smoothing over the key. Should be an integer. If \code{neighbour = x}, \code{x}
number of keys before the key of interest and \code{x} number of keys after the
key of interest are grouped together for model fitting.}

\item{index.vars}{A character vector of names of the predictor variables for
which are estimated.}

\item{index.ind}{An integer vector that assigns group index for each
predictor in \code{index.vars}.}

\item{index.data}{A \code{tibble} including columns for the constructed indices.}

\item{index.names}{A character vector of names of the constructed indices.}

\item{alpha}{A vector of index coefficients.}

\item{s.vars}{A character vector of names of the predictor variables for
which splines should be fitted individually (rather than considering as a
part of an index considered in \code{index.vars}).}

\item{linear.vars}{A character vector of names of the predictor variables
that are included linearly in the model.}
}
\description{
Converts a given object of class \code{gam} to an object of class \code{smimodelFit}.
}
