% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_backward.R
\name{model_backward}
\alias{model_backward}
\title{Nonparametric Additive Model with Backward Elimination}
\usage{
model_backward(
  data,
  val.data,
  yvar,
  family = gaussian(),
  neighbour = 0,
  s.vars = NULL,
  s.basedim = NULL,
  linear.vars = NULL,
  tol = 0.001,
  parallel = FALSE,
  workers = NULL,
  recursive = FALSE,
  recursive_colRange = NULL
)
}
\arguments{
\item{data}{The data set on which the model(s) will be trained. Must be a
data set of class \code{tsibble}.}

\item{val.data}{Validation data set. (The data set on which the penalty
parameter selection will be performed.) Must be a data set of class
\code{tsibble}. (Once the penalty parameter selection is completed, the best
model will be re-fitted for the combined data set \code{data} + \code{val.data}.)}

\item{yvar}{Name of the response variable as a character string.}

\item{family}{A description of the error distribution and link function to be
used in the model (see \code{\link{glm}} and \code{\link{family}}).}

\item{neighbour}{If multiple models are fitted: Number of neighbours of each
key (i.e. grouping variable) to be considered in model fitting to handle
smoothing over the key. Should be an integer. If \code{neighbour = x}, \code{x}
number of keys before the key of interest and \code{x} number of keys after the
key of interest are grouped together for model fitting. The default is \code{0}
(i.e. no neighbours are considered for model fitting).}

\item{s.vars}{A character vector of names of the predictor variables for
which splines should be fitted (i.e. non-linear predictors). (default:
NULL)}

\item{s.basedim}{Dimension of the bases used to represent the smooth terms
corresponding to \code{s.vars}. (For more information refer \code{mgcv::s()}.)
(default: NULL)}

\item{linear.vars}{A character vector of names of the predictor variables
that should be included linearly into the model (i.e. linear predictors).
(default: NULL)}

\item{tol}{Tolerance for the ratio of relative change in MSE, used in model
selection. (default: 0.001)}

\item{parallel}{Whether to use parallel computing in model selection or not.
(default: FALSE)}

\item{workers}{If \code{parallel = TRUE}, number of workers to use. (default:
NULL)}

\item{recursive}{Whether to obtain recursive forecasts or not (default -
FALSE).}

\item{recursive_colRange}{If \code{recursive = TRUE}, The range of column numbers
in \code{val.data} to be filled with forecasts.}
}
\description{
Fits a nonparametric additive model, with simultaneous variable selection
through a backward elimination procedure as proposed by Fan and Hyndman
(2012).
}
\examples{
library(dplyr)
library(tibble)
library(tidyr)
library(tsibble)
n = 1205
set.seed(123)
sim_data <- tibble(x_lag_000 = runif(n)) |>
  mutate(
    # Add x_lags
    x_lag = lag_matrix(x_lag_000, 5)) |>
  unpack(x_lag, names_sep = "_") |>
  mutate(
    # Response variable
    y1 = (0.9*x_lag_000 + 0.6*x_lag_001 + 0.45*x_lag_003)^3 + rnorm(n, sd = 0.1),
    # Add an index to the data set
    inddd = seq(1, n)) |>
  drop_na() |>
  select(inddd, y1, starts_with("x_lag")) |>
  # Make the data set a `tsibble`
  as_tsibble(index = inddd)
# Training set
sim_train <- sim_data[1:1000, ]
# Validation set
sim_val <- sim_data[1001:1200, ]
# Smooth variables
s.vars <- colnames(sim_data)[3:8]
# Model fitting
backwardModel <- model_backward(data = sim_train,
                                val.data = sim_val,
                                yvar = "y1",
                                s.vars = s.vars)
backwardModel$fit[[1]]

}
\references{
Fan, S. & Hyndman, R.J. (2012). Short-Term Load Forecasting Based
on a Semi-Parametric Additive Model. \emph{IEEE Transactions on Power Systems},
27(1), 134-141. \url{http://doi.org/10.1109/TPWRS.2011.2162082}
}
