% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_gaim.R
\name{model_gaim}
\alias{model_gaim}
\title{Groupwise Additive Index Models (GAIM)}
\usage{
model_gaim(
  data,
  yvar,
  neighbour = 0,
  index.vars,
  index.ind,
  s.vars = NULL,
  linear.vars = NULL,
  ...
)
}
\arguments{
\item{data}{Training data set on which models will be trained. Must be a data
set of class \code{tsibble}.(Make sure there are no additional
date/time/date-time/yearmonth/POSIXct/POSIXt variables except for the
\code{index} of the \code{tsibble}). If multiple models are fitted, the grouping
variable should be the key of the \code{tsibble}.}

\item{yvar}{Name of the response variable as a character string.}

\item{neighbour}{If multiple models are fitted: Number of neighbours of each
key (i.e. grouping variable) to be considered in model fitting to handle
smoothing over the key. Should be an integer. If \code{neighbour = x}, \code{x}
number of keys before the key of interest and \code{x} number of keys after the
key of interest are grouped together for model fitting. The default is \code{0}
(i.e. no neighbours are considered for model fitting).}

\item{index.vars}{A character vector of names of the predictor variables for
which indices should be estimated.}

\item{index.ind}{An integer vector that assigns group index for each
predictor in \code{index.vars}.}

\item{s.vars}{A character vector of names of the predictor variables for
which splines should be fitted individually (rather than considering as a
part of an index considered in \code{index.vars}).}

\item{linear.vars}{A character vector of names of the predictor variables
that should be included linearly into the model.}

\item{...}{Other arguments not currently used. (Note that the arguments in
\code{cgaim::cgaim()} related to constrained GAIMs are currently not supported.
Furthermore, the argument \code{subset} is also not supported due to a bug in
\code{cgaim::cgaim()}.)}
}
\description{
A wrapper for \code{cgaim::cgaim()} enabling multiple GAIM models based on a
grouping variable. Currently does not support Constrained GAIM (CGAIM)s.
}
\examples{
library(cgaim)
library(dplyr)
library(tibble)
library(tidyr)
library(tsibble)
n = 1005
set.seed(123)
sim_data <- tibble(x_lag_000 = runif(n)) |>
  mutate(
    # Add x_lags
    x_lag = lag_matrix(x_lag_000, 5)) |>
  unpack(x_lag, names_sep = "_") |>
  mutate(
    # Response variable
    y1 = (0.9*x_lag_000 + 0.6*x_lag_001 + 0.45*x_lag_003)^3 + rnorm(n, sd = 0.1),
    # Add an index to the data set
    inddd = seq(1, n)) |>
  drop_na() |>
  select(inddd, y1, starts_with("x_lag")) |>
  # Make the data set a `tsibble`
  as_tsibble(index = inddd)
# Index variables
index.vars <- colnames(sim_data)[3:7]
# Assign group indices for each predictor
index.ind = c(rep(1, 3), rep(2, 2))
# Smooth variables not entering indices
s.vars = "x_lag_005"
# Model fitting
gaimModel <- model_gaim(data = sim_data,
                        yvar = "y1",
                        index.vars = index.vars,
                        index.ind = index.ind,
                        s.vars = s.vars)
gaimModel$fit[[1]]
}
