% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_groupSmimodel.R
\name{new_groupSmimodel}
\alias{new_groupSmimodel}
\title{Constructor function for the class \code{groupSmimodel}}
\usage{
new_groupSmimodel(
  data,
  yvar,
  family = gaussian(),
  index.vars,
  initialise = c("additive", "linear", "userInput"),
  index.ind = NULL,
  index.coefs = NULL,
  neighbour = 0,
  linear.vars = NULL
)
}
\arguments{
\item{data}{Training data set on which models will be trained. Must be a data
set of class \code{tsibble}.(Make sure there are no additional
date/time/date-time/yearmonth/POSIXct/POSIXt variables except for the
\code{index} of the \code{tsibble}). If multiple models are fitted, the grouping
variable should be the key of the \code{tsibble}.}

\item{yvar}{Name of the response variable as a character string.}

\item{family}{A description of the error distribution and link function to be
used in the model (see \code{\link{glm}} and \code{\link{family}}).}

\item{index.vars}{A character vector of names of the predictor variables for
which indices should be estimated.}

\item{initialise}{The model structure with which the estimation process
should be initialised. The default is "additive", where the initial model
will be a nonparametric additive model. The other options are "linear" -
linear regression model (i.e. a special case single-index model, where the
initial values of the index coefficients are obtained through a linear
regression), and "userInput" - user specifies the initial model structure
(i.e. the number of indices and the placement of index variables among
indices) and the initial index coefficients through \code{index.ind} and
\code{index.coefs} arguments respectively.}

\item{index.ind}{If \code{initialise = "userInput"}: an integer vector that
assigns group index for each predictor in \code{index.vars}.}

\item{index.coefs}{If \code{initialise = "userInput"}: a numeric vector of index
coefficients.}

\item{neighbour}{If multiple models are fitted: Number of neighbours of each
key (i.e. grouping variable) to be considered in model fitting to handle
smoothing over the key. Should be an integer. If \code{neighbour = x}, \code{x}
number of keys before the key of interest and \code{x} number of keys after the
key of interest are grouped together for model fitting. The default is \code{0}
(i.e. no neighbours are considered for model fitting).}

\item{linear.vars}{A character vector of names of the predictor variables
that should be included linearly into the model.}
}
\description{
Constructs an object of class \code{groupsmimodel} (A tibble of Multiple
\code{smimodel}s fitted based on a grouping variable of interest.) using the
information passed to arguments.
}
