% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_smimodel.R
\name{new_smimodel}
\alias{new_smimodel}
\title{Constructor function for the class \code{smimodel}}
\usage{
new_smimodel(
  data,
  yvar,
  index.vars,
  initialise = c("additive", "linear", "userInput"),
  index.ind = NULL,
  index.coefs = NULL,
  linear.vars = NULL
)
}
\arguments{
\item{data}{Training data set on which models will be trained. Should be a
\code{tibble}.}

\item{yvar}{Name of the response variable as a character string.}

\item{index.vars}{A character vector of names of the predictor variables for
which indices should be estimated.}

\item{initialise}{The model structure with which the estimation process
should be initialised. The default is "additive", where the initial model
will be a nonparametric additive model. The other options are "linear" -
linear regression model (i.e. a special case single-index model, where the
initial values of the index coefficients are obtained through a linear
regression), and "userInput" - user specifies the initial model structure (i.e.
the number of indices and the placement of index variables among indices)
and the initial index coefficients through \code{index.ind} and \code{index.coefs}
arguments respectively.}

\item{index.ind}{If \code{initialise = "other"}: an integer vector that assigns
group index for each predictor in \code{index.vars}.}

\item{index.coefs}{If \code{initialise = "other"}: a numeric vector of index
coefficients.}

\item{linear.vars}{A character vector of names of the predictor variables
that should be included linearly into the model.}
}
\description{
Constructs an object of class \code{smimodel} using the information passed to
arguments.
}
