% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_alpha.R
\name{normalise_alpha}
\alias{normalise_alpha}
\title{Scaling index coefficient vectors to have unit norm}
\usage{
normalise_alpha(alpha)
}
\arguments{
\item{alpha}{A vector of index coefficients.}
}
\description{
Scales a coefficient vector of a particular index to have unit norm.
}
