% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\docType{data}
\name{MSE}
\alias{MSE}
\alias{MAE}
\alias{point_measures}
\title{Point estimate accuracy measures}
\format{
An object of class \code{list} of length 2.
}
\usage{
MSE(residuals, na.rm = TRUE, ...)

MAE(residuals, na.rm = TRUE, ...)

point_measures
}
\arguments{
\item{residuals}{A vector of residuals from either the validation or test
data.}

\item{na.rm}{If \code{TRUE}, remove the missing values before calculating the
accuracy measure.}

\item{...}{Additional arguments for each measure.}
}
\description{
Point estimate accuracy measures
}
\keyword{datasets}
