% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb.R
\name{possibleFutures_smimodel}
\alias{possibleFutures_smimodel}
\title{Possible future sample paths (multi-step) from \code{smimodel} residuals}
\usage{
possibleFutures_smimodel(object, newdata, bootstraps, recursive_colRange)
}
\arguments{
\item{object}{A \code{smimodel} object.}

\item{newdata}{The set of new data on for which the forecasts are required
(i.e. test set; should be a \code{tsibble}).}

\item{bootstraps}{Generated matrix of bootstrapped residual series.}

\item{recursive_colRange}{The range of column numbers in \code{newdata} to be
filled with forecasts.}
}
\description{
Generates possible future sample paths (multi-step) using residuals of a
fitted \code{smimodel} through recursive forecasting.
}
