% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.gaimFit}
\alias{print.gaimFit}
\title{Printing a \code{gaimFit} object}
\usage{
\method{print}{gaimFit}(x, ...)
}
\arguments{
\item{x}{A model object of class \code{gaimFit} as produced by \code{model_gaim()}
function.}

\item{...}{Other arguments not currently used.}
}
\description{
The default print method for a \code{gaimFit} object.
}
