% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.smimodel}
\alias{print.smimodel}
\title{Printing a \code{smimodel} object}
\usage{
\method{print}{smimodel}(x, ...)
}
\arguments{
\item{x}{A model object of class \code{smimodel} as produced by \code{model_smimodel()}
function.}

\item{...}{Other arguments not currently used.}
}
\description{
The default print method for a \code{smimodel} object.
}
