% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R, R/autoplot.R, R/model_gaim.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{augment}
\alias{autoplot}
\alias{g}
\alias{s}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{cgaim}{\code{\link[cgaim]{g}}, \code{\link[cgaim:g]{s}}}

  \item{generics}{\code{\link[generics]{augment}}}

  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}
}}

