% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smimodel-package.R
\docType{package}
\name{smimodel-package}
\alias{smimodel}
\alias{smimodel-package}
\title{smimodel: Sparse Multiple Index (SMI) Models for High-dimensional Nonparametric Forecasting}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A general algorithm for estimating Sparse Multiple Index Models for high-dimensional nonparametric forecasting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nuwani-palihawadana/smimodel}
  \item Report bugs at \url{https://github.com/nuwani-palihawadana/smimodel/issues}
}

}
\author{
\strong{Maintainer}: Nuwani Palihawadana \email{Nuwani.KodikaraPalihawadana@monash.edu} (\href{https://orcid.org/0009-0008-6395-7797}{ORCID})

Authors:
\itemize{
  \item Rob Hyndman \email{Rob.Hyndman@monash.edu} (\href{https://orcid.org/0000-0002-2140-5352}{ORCID})
  \item Xiaoqian Wang \email{Xiaoqian.Wang@monash.edu} (\href{https://orcid.org/0000-0003-4827-496X}{ORCID})
}

}
\keyword{package}
