% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smimodel-package.R
\docType{package}
\name{smimodel-package}
\alias{smimodel}
\alias{smimodel-package}
\title{smimodel: Sparse Multiple Index (SMI) Models for Nonparametric Forecasting and Prediction}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Implements a general algorithm for estimating Sparse Multiple Index (SMI) models for nonparametric forecasting and prediction. Includes functions for fitting nonparametric additive models with backward elimination, group-wise additive index models, and projection pursuit regression models as benchmark comparison methods. Provides tools for generating prediction intervals to quantify uncertainty in point forecasts produced by the SMI model and benchmark models, using either the classical Block Bootstrap method or a novel Conformal Bootstrap approach that integrates block bootstrap with split conformal prediction through an online updating procedure.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nuwani-palihawadana/smimodel}
  \item \url{https://nuwani-palihawadana.github.io/smimodel/}
  \item Report bugs at \url{https://github.com/nuwani-palihawadana/smimodel/issues}
}

}
\author{
\strong{Maintainer}: Nuwani Palihawadana \email{Nuwani.KodikaraPalihawadana@monash.edu} (\href{https://orcid.org/0009-0008-6395-7797}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Xiaoqian Wang \email{Xiaoqian.Wang@amss.ac.cn} (\href{https://orcid.org/0000-0003-4827-496X}{ORCID}) [contributor]
}

}
\keyword{package}
