% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smimodel.R
\name{smimodel}
\alias{smimodel}
\title{Sparse Multiple Index (SMI) models}
\usage{
smimodel(
  data,
  yvar,
  index.vars,
  initialise = c("additive", "linear", "userInput"),
  index.ind = NULL,
  index.coefs = NULL,
  linear.vars = NULL,
  lambda0 = 1,
  lambda2 = 1,
  M = 10,
  max.iter = 50,
  tol = 0.001,
  tolCoefs = 0.001,
  TimeLimit = Inf,
  verbose = FALSE
)
}
\arguments{
\item{data}{Training data set on which models will be trained. Should be a
\code{tibble}.}

\item{yvar}{Name of the response variable as a character string.}

\item{index.vars}{A character vector of names of the predictor variables for
which indices should be estimated.}

\item{initialise}{The model structure with which the estimation process
should be initialised. The default is "additive", where the initial model
will be a nonparametric additive model. The other options are "linear" -
linear regression model (i.e. a special case single-index model, where the
initial values of the index coefficients are obtained through a linear
regression), and "userInput" - user specifies the initial model structure (i.e.
the number of indices and the placement of index variables among indices)
and the initial index coefficients through \code{index.ind} and \code{index.coefs}
arguments respectively.}

\item{index.ind}{If \code{initialise = "userInput"}: an integer vector that assigns
group index for each predictor in \code{index.vars}.}

\item{index.coefs}{If \code{initialise = "userInput"}: a numeric vector of index
coefficients.}

\item{linear.vars}{A character vector of names of the predictor variables
that should be included linearly into the model.}

\item{lambda0}{Penalty parameter for L0 penalty.}

\item{lambda2}{Penalty parameter for L2 penalty.}

\item{M}{Big-M value used in MIP.}

\item{max.iter}{Maximum number of MIP iterations performed to update index
coefficients for a given model.}

\item{tol}{Tolerance for loss.}

\item{tolCoefs}{Tolerance for coefficients.}

\item{TimeLimit}{A limit for the total time (in seconds) expended in a single
MIP iteration.}

\item{verbose}{The option to print detailed solver output.}
}
\description{
Fits a nonparametric multiple index model to data with simultaneous variable
selection (hence "sparse").
}
