% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_smimodel.R
\name{split_index}
\alias{split_index}
\title{Splitting predictors into multiple indices}
\usage{
split_index(num_pred, num_ind)
}
\arguments{
\item{num_pred}{Number of predictors.}

\item{num_ind}{Number of indices.}
}
\value{
A list containing the following components: \item{index}{An integer
vector that assigns group indices for each predictor.}
\item{index_positions}{A list of length = \code{num_ind} that indicates
which predictors belong to which index.}
}
\description{
Splits a given number of predictors into a given number of indices.
}
