% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{truncate_vars}
\alias{truncate_vars}
\title{Truncating predictors to be in the in-sample range}
\usage{
truncate_vars(object.data, data, cols.trunc)
}
\arguments{
\item{object.data}{In-sample data set.}

\item{data}{Out-of-sample data set of which variables should be truncated.}

\item{cols.trunc}{Column names of the variables to be truncated.}
}
\description{
Truncates predictors to be in the in-sample range to avoid spline
extrapolation.
}
